fix: split relay registry tests
